workaround CMake heavy caching behaviour
CMake modules have a tendency to cache way too much. The net
effect is that changes that are detected by autobuild are
actually not properly propagated during the build because
CMake reuses old values.

The default behaviour for autobuild is now to delete the
CMakeCache.txt file whenever a reconfiguration is needed. This
behaviour can be configured system-wide with
Autobuild::CMake.full_reconfigures= and per-package with the
corresponding instance method.