fix the orogen package handler in bootstrapping contexts
The issue is that the orogen package handler was assuming
that "orogen" is available, which may not be true at bootstrapping

It is a nicety -- i.e. not very required. Moreover, that most
of the time "orogen" not being available means that we *are*
bootstrapping. So silently ignore orogen not being there.