fix handling of per-package log directories
That was completely broken, as Subcommand.run was always using
the global logdir. The first argument of Subcommand.run can now
either be a string or a package object. In the first case, the
global log dir will be used while in the second case the value
returned by #logdir will.