we must not use installstamp in #depends_on
The issue is that the path to #installstamp depends on the
values of #srcdir, #logdir and #prefix. These values may be
set *after* the package spec has been loaded, *before* the build

So, move any additional dependency that relies on #installstamp
to #prepare, where the *dir values are guaranteed to be properly
set.