make the environment configuration easier for the user
the #env_add_path method was filtering the paths that did not
exist yet. This is harmful for users of the autobuil API, as they
have to make sure that the provided path is already created before
it can be added to the environment.

Assume that the caller knows what it is doing, and just add it.

Also, create the Package#update_environment method that is supposed
to update the environment. This method gets called at import and
install time.