autotools: make the checkout detection robust to build interruptions
Autobuild forces a regen after the first checkout. The issue is that
autotools is less robust than it should, and very often it is
better to generate the build system for the system on which we
must build, to avoid broken builds.

The current method was saving a local variable in the package object,
which was telling #prepare to do a forced build. This is obviously
not robust to interruptions that could happen between #import and
#prepare.

This commits replaces the attribute a file touched in the
source directory, when we do a fresh checkout.