git: greatly improve the handling of commit pinning
The code now tries to stay on proper branches, instead of creating
detached HEADs. Moreover, will also avoid to do non-simple updates
even if specific commits are specified, as it could make people
lose some data if they were developping on the detached HEAD and switch
to another commit.

Obviously, this check is not present when you check it out.