fix importer updates with patching when patches add new files
This issue is related to 44105115460aaf9861550e6c7bc22ff9faa4c35e

The problem that 44105115460aaf9861550e6c7bc22ff9faa4c35e was supposed
to fix for zip importers is actually not fixed. The issue was not the
behaviour of unzip (the -o option ensured that unzip had the right
behaviour), but the fact that the failing patch was creating a new file,
something that unzip obviously does not handle.

Add the --forward option to the patch command so that already applied
chunks are simply ignored.