implement parallel-build management at the package level
This commit implements the ability to build packages in parallel
if they don't depend on each other. Moreover, the "jobserver"
implementation is compatible with GNU make, which makes it possible
to dynamically distribute jobs across different packages that use
make (a very common thing).

It required reworking the messaging API. In particular, single messages
are now sent with Autobuild.message (was Autobuild.progress) and build
progress messages must be sent with the Autobuild.progress_start /
Autobuild.progress and Autobuild.progress_done set of calls (and the
corresponding versions on Autobuild::Package)