doc generation conflicts with build during parallel build processing, force ordering
Documentation generation does not strictly need the package to be built and
installed. However, for some packages, they do require their dependencies to be
built and installed.

In a parallel build setting, it means that we might end up with a build-related
target and the doc target being executed at the same time, which fails for
concurrency reasons in a lot of cases.

For now, force the ordering between build and documentation. Later, this
particular issue will be fixed by ensuring, during the parallel build
processing, that two tasks from the same package cannot be executed in parallel.