fix parallel builds on ruby 2.0
The behaviour of IOs and exec() changed between 1.9 and 2.0. Namely,
all IOs have the close-on-exec flag set, and exec() is also set to
close all non-standard FDs automatically.

This is breaking the interaction between the job server and gnumake
since this requires a pipe to be shared between autobuild and the
gnumake processes