split identification of importers between source_id and repository_id
repository_id is meant to identify the repository, i.e. where the
data is, regardless of e.g. access protocol

source_id is meant to identify what the importer points to,
regardless of configuration options that have no effect on the data
itself (e.g. git would include the branch but remove the retry_count)