fix race condition in parallel processing
task.already_invoked is set to true by Rake *before* the task
is actually executed. In some cases, it was causing tasks whose
prerequisites were being executed to be triggered. Since the
task was being executed but was not finished, Rake would re-trigger
the prerequisite in parallel to the first one and (usually) boom.