git: never create a detached HEAD
The current behaviour when commit: or tag: was given was to create
a detached HEAD, which is really really less than ideal.

Since we are going to use those a lot more, this commit tries to find
a better way. Namely, autobuild now resets the local branch to the
commit while trying to preserve uncommitted changes. It will do so
only if the branch contains only commits that are already on the remote
origin.