more fine-grained modularized parallel make support and rake integration
This is meant to allow parallelizing make-based builds that are
"hidden" under rake. The only things that one would need to do
is override invoke_rake to tune rake_setup_task using the parallel
build commands given to it, e.g. assuming that the 'build' task
accepts the make arguments,

    def pkg.invoke_rake
	Autobuild.invoke_make_parallel do |*options|
	    self.rake_setup_task = "build[#{options.join(" ")}]"
	    super
	end
    end