give control to enable/disable utilities freely
This commit is meant to give the means to autoproj to enable/disable
utilities (here, tests) on a package-by-package basis. This requires
to split the availability of utilities (whether tests are present at
all) with the control over enabling/disabling them.

It then adds an option to declare whether an utility should be
enabled by default or not, and disables tests by default