fix SVN#svn_info when called with verbose == true
Calling #run with a block is meant to be used for progress display only.
So, when verbose is true, the output is NOT yield to the caller (as
progress would duplicate the actual program output)

Use the return value of #run, and provide a new way to get the output
in the event of a failure.

Conflicts:
	lib/autobuild/subcommand.rb