importers: implement the reset flag
This closes rock-core/autoproj#39 and is required for the
implementation of rock-core/autoproj#37.

It adds a reset option to #import which controls what to do if
the local working copy is not exactly the state specified by
the importer configuration. The only one which acts on it is
really the git importer, the other ones simply ignore it.

This modifies the importer API in a backward-compatible way, changing
the second argument from a boolean to an option hash, but allowing the
boolean to be still passed as-is

In addition, it implements a fast path in the git importer in case
the commit/tag is already available.