git: fix mixup between tags and branches when fetching remotes
The semantic of the default refspec in fetch_remote and
current_remote_commit was really unclear. One would assume that they
fetch the remote tag if that's what the importer needs, but they
were actually fetching the remote branch always, for the benefit of
e.g. #update which needs it in case of resets.

This commit makes #fetch_remote explicitely require the caller to
pass a remote refspec. #current_remote_commit now resolves the tag
if required, and then fallbacks to the remote by default, but also
accepts a refspec argument. #update is modified to pass the refspec
explicitely for its own needs.