do not filter InteractionRequired exceptions in isolate_errors
It is a coding error, not an import error. Either the caller should
be able to handle it, or should not have called the method at all.

It fixes the handling of sometimes-interactive packages in autoproj 2
(as e.g. archives that need to be updated). Without this change, calling
autoproj with -k would cause the whole update process to error.