'modernize' the orogen package handler
It does not access the ENV hash anymore, but uses its own package-local
environment to find orogen. Moreover, it does not attempt to autodetect
its dependency on orogen/rtt in weird ways, instead assuming that the
layer on top (e.g. autoproj) sets dependencies up properly.

It's also stricter with finding orogen, and will raise if it cannot find
it and/or cannot guess the orogen version