git-lfs: add support for include/exclude patterns
This commit allows to set a lfs_include/lfs_exclude options in the
importer configuration to control git-lfs's corresponding feature.
The two options need to be exactly what's expected by git-lfs, which
is mainly a comma-separated list of glob patterns. See git-lfs-fetch(1)
for more details.

Moreover, git-lfs is now unconditionally setup as soon as a 'lfs'
directory is found in .git/. What is controlled by the 'lfs' key is
solely whether autobuild should pull the lfs files or not. This ensures
that the repository is usable, e.g. that one can cd into it and run
git-lfs commands manually.