fix default setup in CMake#common_utility_handling
It was setting source_dir to package.srcdir, but source_dir in a
utility is the intermediate directory in which the utility generates
its files (target_dir being where they should be installed).

Set source_ref_dir to builddir, since we assume that the utility
byproducts be in builddir/{utility}, e.g.  builddir/doc