use Ruby's built-in Net::HTTP to deal with archives from HTTP
This adds authentication and progress support, and makes the
importer more robust to update detection. Moreover, it's
cross-platform, which ensures that the same code path is taken
on Windows and Linux/BSD/MacOS. Currently, we were using a
pure-Ruby codepath on Windows, curl on MacOS/BSD and wget on Linux.