do process dependencies of disabled tasks
Until now, #disabled! would mark tasks as already invoked and
not needed, which in turn is meant to stop processing from the
perspective of Rake's original processing logic, which we
duplicated.

The downside of this method is that it really stops the graph
search. So, disabling a package that is in the middle of the
dependency search breaks dependency-based ordering. This is why
when running `amake --deps=f` with a set of packages, one often
gets compilation in the wrong order.

However, Autobuild now has its own task execution algorithm.
Moreover, tasks within Autobuild are associated to a package.

This commit uses both these facts to (1) only mark a package
as disabled, and have the tasks associated with the package
be disabled as well. This ensures that all tasks related to
a given package are disabled in one shot and (2) look for
disabled? explicitly in the resolution algorithm to avoid
processing a task, but still looking to execute its dependencies.