feat(git): allow to fingerprint on tree instead of commit
The main feature of fingerprints is to figure out whether the
source code has changed, for e.g. CI purposes. Using the tree
avoids to re-build things for which the commit changed but the
source code actually did not.

This obviously only works if there are no checkout filters that
modify the files between the commit object and the working copy.

Since changing it could have an impact, I preferred keeping the
old behavior as default, while allowing to override globally
through Git.default_fingerprint_mode= and locally as an argument
to the importer.