fix: do not depend on git-lfs being installed during git clone
The current git-lfs integration runs `git lfs pull` after the normal
git operations. We therefore setup the git repositories to essentially
do nothing on clone/fetch/pull/...

However, the way this is done so far is by using the `--skip` options
to git-lfs ... which relies on git-lfs itself. This is a problem
on platforms where installing git-lfs is an issue.

Use pass-through filters that do not rely on git-lfs during cloning
instead. The "real" setup is done after the import anyways.